requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-cors-headers==4.3.1
django-split-settings==1.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1

# Production dependencies
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
